// AngularUi/BuaAndSellGold/src/app/models/goldPrice.interface.ts
export interface goldPrice2 {

    "gold": [
        {
            "name": "انس جهانی طلا",
            "price": 2620
        },
        {
            "name": "مثقال طلا",
            "price": 21093000
        },
        {
            "name": "گرم طلای 18 عیار",
            "price": 4869340
        },
        {
            "name": "گرم طلای 24 عیار",
            "price": 6478600
        },
        {
            "name": "سکه بهار آزادی",
            "price": 51600000
        },
        {
            "name": "سکه امامی",
            "price": 53600000
        },
        {
            "name": "نیم سکه",
            "price": 29000000
        },
        {
            "name": "ربع سکه",
            "price": 17200000
        },
        {
            "name": "سکه گرمی",
            "price": 8200000
        }
    ],
    "currency": [
        {
            "name": "دلار",
            "price": 76100
        },
        {
            "name": "یورو",
            "price": 79350
        },
        {
            "name": "درهم امارات",
            "price": 20920
        },
        {
            "name": "پوند",
            "price": 95650
        },
        {
            "name": "کرون سوئد",
            "price": 6895
        },
        {
            "name": "کرون دانمارک",
            "price": 10650
        },
        {
            "name": "صد ین ژاپن",
            "price": 4865
        },
        {
            "name": "رینگیت مالزی",
            "price": 16900
        },
        {
            "name": "دینار کویت",
            "price": 247100
        },
        {
            "name": "لیر سوریه",
            "price": 30
        },
        {
            "name": "روبل روسیه",
            "price": 740
        },
        {
            "name": "ریال عمان",
            "price": 197750
        },
        {
            "name": "دلار سنگاپور",
            "price": 56100
        },
        {
            "name": "دینار عراق",
            "price": 5830
        },
        {
            "name": "افغانی",
            "price": 1096
        },
        {
            "name": "منات آذربایجان",
            "price": 44650
        },
        {
            "name": "دلار استرالیا",
            "price": 47600
        },
        {
            "name": "دلار کانادا",
            "price": 53050
        },
        {
            "name": "یوان چین",
            "price": 10450
        },
        {
            "name": "فرانک سوئیس",
            "price": 85150
        },
        {
            "name": "کرون نروژ",
            "price": 6720
        },
        {
            "name": "لیر ترکیه",
            "price": 2165
        },
        {
            "name": "دلار نیوزیلند",
            "price": 43060
        },
        {
            "name": "دلار هنگ کنگ",
            "price": 9785
        },
        {
            "name": "ریال عربستان",
            "price": 20250
        },
        {
            "name": "روپیه هند",
            "price": 895
        },
        {
            "name": "روپیه پاکستان",
            "price": 275
        },
        {
            "name": "دینار بحرین",
            "price": 203590
        },
        {
            "name": "ریال قطر",
            "price": 20950
        },
        {
            "name": "بات تایلند",
            "price": 2225
        },
        {
            "name": "درام ارمنستان",
            "price": 1935
        },
        {
            "name": "لاری گرجستان",
            "price": 27310
        }
    ],
    "cryptocurrency": [
        {
            "name": "بیت کوین",
            "price": 96813
        },
        {
            "name": "اتریوم",
            "price": 3367
        }
    ]

}